PROBLEM FIX
after fixing minor issue with library i was using to retrieve book data, i realised i could not get the book genre from any API with the book ISBN. this means that i will have to wtrite my own solution.
I have decided to create a book genres table where genres can be added/removed, and i will use the wikipedia list of genres to make this table. The genre will be selected on the creation of the book.