PROBLEM FIX
global vars reset whenever i switched pages. I decided to use cookies to store my global variables such as the UserID that will be used to interact with the database, as well as lock pages when the user is not logged in. I used js-cookies instead of react-cookies becauase react cookies can only be accessed from function components, while i need class components.

also setup all navbar routes to point to each element of the page.

rewrote index.js to render the form if and only if the user is not signed it, otherwise it renders the WelcomePage element.